Multiple constraints updates & addressing warnings.
- Icon bundles upgraded to match required sizes.
- Moving to NSAttributes instead of deprecated UIAttributes.
- Casting serialized attachmentID to NSNumber.
- Removing unused archival view from InboxTableViewCell.
- Upgrading contraints on SignalsVC for 6+ "TabBar".
- Removing deprecated "plain" properties
- Removing redundant button in FingerprintViewController.
- Addresses issues with placement of change number in signup flow
- Fixes placement bug in new group view
- Exclude unused reference.

+++ Signal/src/view controllers/MessagesViewController.m
@@ -1109,8 +1109,8 @@ typedef enum : NSUInteger {
     }
 
     NSURL *compressedVideoUrl = [NSURL fileURLWithPath:basePath];
-    long currentTime = [[NSDate date] timeIntervalSince1970];
-    NSString *strImageName = [NSString stringWithFormat:@"%ld",currentTime];
+    double currentTime = [[NSDate date] timeIntervalSince1970];
+    NSString *strImageName = [NSString stringWithFormat:@"%f",currentTime];
     compressedVideoUrl=[compressedVideoUrl URLByAppendingPathComponent:[NSString stringWithFormat:@"%@.mp4",strImageName]];
 
     exportSession.outputURL = compressedVideoUrl;
